# predicates/host: add HostAny benchmark